Reduce main app bundle by loading less Storyblok components on CMS pages
