+++ apps/store/src/features/blog/BlogContextProvider.tsx
@@ -5,6 +5,7 @@ import type {
   BLOG_ARTICLE_TEASERS_PAGE_PROP,
 } from '@/features/blog/blog.constants'
 import type { BlogArticleCategory, BlogArticleTeaser } from '@/features/blog/blog.types'
+import { BlogStoryblokProvider } from '@/features/blog/BlogStoryblokProvider'
 import { BlogContext, parseBlogContext } from '@/features/blog/useBlog'
 
 type Props = {
@@ -18,9 +19,11 @@ export type BlogPageProps = {
 }
 
 export function BlogContextProvider({ blogPageProps, children }: Props) {
+  const hasBlogContext = Object.values(blogPageProps ?? {}).some(Boolean)
+
   return (
     <BlogContext.Provider value={parseBlogContext(blogPageProps ?? {})}>
-      {children}
+      {hasBlogContext ? <BlogStoryblokProvider>{children}</BlogStoryblokProvider> : children}
     </BlogContext.Provider>
   )
 }
